+++ src/service/placementStrategy/rowCropped/flat/RowCroppedFlatThreeProductImagesPlacementStrategy.php
@@ -43,7 +43,7 @@ class RowCroppedFlatThreeProductImagesPlacementStrategy implements ThreeProductI
     {
         $categoryImage = ImageUtils::createTransparentImage(self::$WIDTH, self::$HEIGHT);
 
-        $productImageSize = (self::$WIDTH / 3) - (2 * RowCroppedConstants::PADDING_BETWEEN);
+        $productImageSize = (int) ((self::$WIDTH / 3) - (2 * RowCroppedConstants::PADDING_BETWEEN));
         $productImage1 = ImageUtils::resizeImageToMaxWidthHeight($productImage1, $productImageSize, $productImageSize);
         $productImage2 = ImageUtils::resizeImageToMaxWidthHeight($productImage2, $productImageSize, $productImageSize);
         $productImage3 = ImageUtils::resizeImageToMaxWidthHeight($productImage3, $productImageSize, $productImageSize);
