+++ src/service/placementStrategy/rowCropped/flat/RowCroppedFlatTwoProductImagesPlacementStrategy.php
@@ -42,7 +42,7 @@ class RowCroppedFlatTwoProductImagesPlacementStrategy implements TwoProductImage
     {
         $categoryImage = ImageUtils::createTransparentImage(self::$WIDTH, self::$HEIGHT);
 
-        $productImageSize = (self::$WIDTH / 3) - (2 * RowCroppedConstants::PADDING_BETWEEN);
+        $productImageSize = (int) ((self::$WIDTH / 3) - (2 * RowCroppedConstants::PADDING_BETWEEN));
         $productImage1 = ImageUtils::resizeImageToMaxWidthHeight($productImage1, $productImageSize, $productImageSize);
         $productImage2 = ImageUtils::resizeImageToMaxWidthHeight($productImage2, $productImageSize, $productImageSize);
